make timer cleaner

+++ framework/organizer_scene/src/mm/services/organizer_scene.hpp
@@ -14,7 +14,7 @@ namespace MM::Services {
 			std::unique_ptr<Scene> _next_scene; // enqueued next scene
 
 			using clock = std::chrono::high_resolution_clock;
-			long long int _accumulator = 0;
+			double _accumulator = 0.0;
 			std::chrono::time_point<clock> _last_time;
 
 		public:
